Medikal Estetik/Profiloplasti/requirements.txt
mediapipe>=0.8.10
opencv-python>=4.7.0
numpy>=1.22.0
Pillow>=9.0.0
reportlab>=3.6.0